# Conflicts:
#	fonts/tmp/Bravura-bounding-boxes.svg
#	fonts/tmp/Gootville-bounding-boxes.svg
#	fonts/tmp/Leipzig-bounding-boxes.svg
#	fonts/tmp/Petaluma-bounding-boxes.svg

+++ libmei/atts_shared.h
@@ -2908,8 +2908,8 @@ public:
      * to the default value)
      **/
     ///@{
-    void SetMm(int mm_) { m_mm = mm_; }
-    int GetMm() const { return m_mm; }
+    void SetMm(double mm_) { m_mm = mm_; }
+    double GetMm() const { return m_mm; }
     bool HasMm() const;
     //
     void SetMmUnit(data_DURATION mmUnit_) { m_mmUnit = mmUnit_; }
@@ -2929,7 +2929,7 @@ private:
      * In MIDI, a beat is always defined as a quarter note, *not the numerator of the
      * time signature or the metronomic indication*.
      **/
-    int m_mm;
+    double m_mm;
     /** Captures the metronomic unit. **/
     data_DURATION m_mmUnit;
     /** Records the number of augmentation dots required by a dotted metronome unit. **/
